Strengthen integration spec for lesson title display
Reduce spec dependence on the implementation of lesson title display in the view and controller. Former specs would have passed if only one lesson had its title displayed out of many lessons in the database. Currently, if there are two lessons in the database, they both must be displayed in order for the specs to pass. To my understanding, all lesson titles will be displayed on the home page for any number of lessons in the database.