Add better test coverage for login feature
Previously, login functionality was only tested through integration tests that checked which path the user was redirected to. This incorrectly assumed that the login state was necessarily coupled with the redirect path, as implemented in the Sessions controller. Now, there are two more controller tests that explicitly test whether the user is logged in, regardless of where the user is redirected to.