Modify integration spec for account activations
The former spec relied on verifying login status via access to the Rails session hash, which is not available in Capybara feature specs. Now, instead, we verify login status for activated and inactivated users by checking for text content which indicates whether the account is activated or not. A major problem with this approach is that the testing strategy is coupled with the implementation (i.e. we have to remember to set up the flash messages in the activations controller to accurately reflect activation status). In the interest of expediency, I am moving forward with this spec anyway, although it is a good candidate for future improvement by checking for login status separately from the implementation of flash messages in the controller. Another set of changes in this spec is two additional expectations (Expect an inactivated user to be at the login page after trying to log in. Expect to be at the login page after following an invalid activation link)