Re-write an unclear spec for better readability
There was a spec that used some hacky error-handling logic to continue running the specs even when an error was raised. This error being raised was part of the specified behavior anyway. So, now, the spec explicitly specifies that an ActiveRecord::RecordInvalid error should be raised when the user attempts to save a duplicate lesson completion object. This approach is more readable, and it brings test coverage up to a clean 100%.